[BAPL-1893] Load kjar before container creation
