+++ pom.xml
@@ -15,7 +15,7 @@
     <maven.compiler.source>11</maven.compiler.source>
     <maven.compiler.testTarget>${maven.compiler.target}</maven.compiler.testTarget>
     <maven.compiler.testSource>${maven.compiler.source}</maven.compiler.testSource>
-    <quarkus.version>2.0.3.Final</quarkus.version>
+    <quarkus.version>2.1.0.CR1</quarkus.version>
     <kieserver.version>${project.version}</kieserver.version>
     <surefire-plugin.version>3.0.0-M5</surefire-plugin.version>
     <lombok.version>1.18.16</lombok.version>
